fixed loginPage issues
