--- frontend/src/components/componentName/styles/Image.styled.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledImage = styled.img`
-  height: auto;
-  width: auto;
-  max-width: ${({ maxW }) => maxW};
-  max-height: ${({ maxH }) => maxH};
-`
