--- frontend/src/components/componentName/styles/Link.styled.js
@@ -1,30 +1,0 @@
-import styled from 'styled-components'
-import { Link } from 'react-router-dom'
-
-export const StyledLink = styled(Link)`
-  // width: 150px;
-  // height: 15px;
-
-  // font-family: Poppins;
-  // font-style: normal;
-  // font-weight: 500;
-  // font-size: 15px;
-  // line-height: 22px;
-
-  // color: #011338;
-`
-
-export const StyledFake = styled.p`
-  height: 15px;
-
-  font-family: Poppins;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 15px;
-  line-height: 22px;
-  cursor: pointer;
-
-  text-align: right;
-
-  color: #011338;
-`
